+++ database/migrations/2022_05_08_065846_create_product_vendor_order_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->enum('product_type', ['simple', 'variable', 'digital', 'service']);
             $table->integer('vendor_order_id');
             $table->integer('warehouse_id');
-            $table->double('product_price', 8, 2)->default(0);
-            $table->double('commission', 8, 2)->default(0);
-            $table->integer('qty')->default(0);
+            $table->double('product_price', 12, 2)->default(0);
+            $table->double('commission', 12, 2)->default(0);
+            $table->double('qty', 12, 2)->default(0);
             $table->timestamps();
         });
     }
